worker/requirements.txt
celery==5.3.6
redis==5.0.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1